% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_helpers.R
\name{apply_config}
\alias{apply_config}
\title{Configuration Helper Functions for Geyser}
\usage{
apply_config(config, session, rv)
}
\arguments{
\item{config}{List containing configuration parameters}

\item{session}{Shiny session object}

\item{rv}{reactiveValues object containing app state}
}
\description{
Functions to handle loading, saving, and applying configurations

Apply a configuration to the Shiny session
}
