% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_helpers.R
\name{apply_embedded_config}
\alias{apply_embedded_config}
\title{Apply embedded config from SummarizedExperiment if present}
\usage{
apply_embedded_config(rse, session, rv)
}
\arguments{
\item{rse}{SummarizedExperiment object}

\item{session}{Shiny session object}

\item{rv}{reactiveValues object}
}
\description{
Apply embedded config from SummarizedExperiment if present
}
