% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_helpers.R
\name{apply_pending_group_filters}
\alias{apply_pending_group_filters}
\title{Apply pending group filters to dynamic UI}
\usage{
apply_pending_group_filters(rv, input, session)
}
\arguments{
\item{rv}{reactiveValues object}

\item{input}{Shiny input object}

\item{session}{Shiny session object}
}
\description{
Apply pending group filters to dynamic UI
}
