% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_helpers.R
\name{apply_pending_sample_filter}
\alias{apply_pending_sample_filter}
\title{Apply pending sample filter to table}
\usage{
apply_pending_sample_filter(rv, input)
}
\arguments{
\item{rv}{reactiveValues object}

\item{input}{Shiny input object}
}
\description{
Apply pending sample filter to table
}
