% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_helpers.R
\name{create_config_download_handler}
\alias{create_config_download_handler}
\title{Create configuration download handler}
\usage{
create_config_download_handler(input)
}
\arguments{
\item{input}{Shiny input object}
}
\description{
Create configuration download handler
}
