% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataApp.R
\name{dataApp}
\alias{dataApp}
\alias{dataServer}
\alias{dataInput}
\alias{dataOutput}
\title{Shiny App for Geyser Graphics Histogram}
\usage{
dataApp()

dataServer(id)

dataInput(id)

dataOutput(id)
}
\arguments{
\item{id}{shiny identifier}
}
\description{
Shiny App for Geyser Graphics Histogram
}
