% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scgeyser_module_data_loader.R
\name{dataLoaderOutputUI}
\alias{dataLoaderOutputUI}
\title{UI for the Data Loader Output Area (Main Panel)}
\usage{
dataLoaderOutputUI(id)
}
\arguments{
\item{id}{Namespace ID.}
}
\description{
Creates the UI for the data loader's output, which appears in the main panel.
}
