% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data_loader.R
\name{dataLoaderServer}
\alias{dataLoaderServer}
\title{Server for the Data Loader Module}
\usage{
dataLoaderServer(id)
}
\arguments{
\item{id}{Namespace ID.}
}
\value{
A reactive list containing `$config`, `$obs_data`, `$gene_table`, and a data access function `$get_gene_data`.
}
\description{
Server for the Data Loader Module
}
