% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{dataServer}
\alias{dataServer}
\alias{dataInput}
\alias{dataOutput}
\alias{dataApp}
\title{Shiny Server for Geyser Graphics Histogram}
\usage{
dataServer(id)

dataInput(id)

dataOutput(id)

dataApp()
}
\arguments{
\item{id}{shiny identifier}
}
\value{
reactive server

nothing returned

nothing returned

nothing returned
}
\description{
Shiny Server for Geyser Graphics Histogram

Shiny Module Input

Shiny Module Output

Shiny Module App
}
