% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsApp.R
\name{datasetsApp}
\alias{datasetsApp}
\alias{datasetsServer}
\alias{datasetsInput}
\alias{datasetsUI}
\alias{datasetsOutput}
\title{Shiny App for Selected R Datasets}
\usage{
datasetsApp()

datasetsServer(id)

datasetsInput(id)

datasetsUI(id)

datasetsOutput(id)
}
\arguments{
\item{id}{shiny identifier}
}
\description{
Shiny App for Selected R Datasets
}
