% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{datasetsServer}
\alias{datasetsServer}
\alias{datasetsInput}
\alias{datasetsOutput}
\alias{datasetsApp}
\title{Shiny Server for Selected R Datasets}
\usage{
datasetsServer(id)

datasetsInput(id)

datasetsOutput(id)

datasetsApp()
}
\arguments{
\item{id}{shiny identifier}
}
\value{
reactive server

nothing returned

nothing returned

nothing returned
}
\description{
Shiny Server for Selected R Datasets

Shiny Module Input

Shiny Module Output

Shiny Module App
}
