% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_dot_plot.R
\name{dotPlotControlsUI}
\alias{dotPlotControlsUI}
\title{UI for Dot Plot Controls}
\usage{
dotPlotControlsUI(id)
}
\arguments{
\item{id}{Namespace ID.}
}
\description{
Creates the UI for the dot plot controls in the sidebar.
}
