% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scgeyser_module_dot_plot.R
\name{dotPlotServer}
\alias{dotPlotServer}
\title{Server for the Dot Plot Module.}
\usage{
dotPlotServer(id, loaded_data)
}
\arguments{
\item{id}{A namespace ID.}

\item{loaded_data}{A reactive list from the dataLoaderServer module.}
}
\description{
Handles the logic for generating and displaying a dot plot based on user selections.
}
