% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_plot.R
\name{exp_plot}
\alias{exp_plot}
\title{exp_plot}
\usage{
exp_plot(input, rse_name, slot)
}
\arguments{
\item{input}{From ui.R}

\item{rse_name}{Name of the rse object}

\item{slot}{which slot to pull the count data from the rse assay}
}
\description{
draws the expression box plot
}
\details{
Makes the box plot for the geyser Shiny app
}
\examples{

\dontrun{
exp_plot(input, rse_name, 'counts')
}

}
\author{
David McGaughey
}
\keyword{internal}
