% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exPlotR.R
\name{exPlotR}
\alias{exPlotR}
\title{exPlotR}
\usage{
exPlotR(rse, ...)
}
\arguments{
\item{rse}{SummarizedExperiment object}

\item{...}{More arguments for \code{\link[shiny:runApp]{shiny::runApp()}}.}
}
\description{
Run shiny app to use SummarizedExperiment object to display RNAseq data
}
\details{
Shiny app uses the rowData rownames to define the genes. The colData field is made
fully available to make custom plot groupings.
}
\examples{

\dontrun{
exPlotR(your_rse)
}

}
\author{
David McGaughey
}
