% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_expression_plot.R
\name{expressionPlotControlsUI}
\alias{expressionPlotControlsUI}
\title{UI for Expression Plot Controls}
\usage{
expressionPlotControlsUI(id)
}
\arguments{
\item{id}{Namespace ID.}
}
\description{
Creates the UI for the expression plot controls in the sidebar.
}
