% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_expression_plot.R
\name{expressionPlotOutputUI}
\alias{expressionPlotOutputUI}
\title{UI for Expression Plot Output Area}
\usage{
expressionPlotOutputUI(id)
}
\arguments{
\item{id}{Namespace ID.}
}
\description{
UI for Expression Plot Output Area
}
