% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scgeyser_module_expression_plot.R
\name{expressionPlotServer}
\alias{expressionPlotServer}
\title{Server for the Expression Plot Module}
\usage{
expressionPlotServer(id, loaded_data)
}
\arguments{
\item{id}{Namespace ID.}

\item{loaded_data}{A reactive list from the dataLoaderServer module.}
}
\description{
Handles the logic for generating and displaying violin or box plots.
}
