% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_gene_table.R
\name{geneTableServer}
\alias{geneTableServer}
\title{Server for the Gene Table Module}
\usage{
geneTableServer(id, gene_data)
}
\arguments{
\item{id}{Namespace ID.}

\item{gene_data}{A reactive data.table of gene information.}
}
\value{
A reactive containing the name of the gene selected from the table.
}
\description{
A reusable server component for the gene table that returns the selected gene.
}
