% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scgeyser_module_gene_table.R
\name{geneTableUI}
\alias{geneTableUI}
\title{UI for the Gene Table Module}
\usage{
geneTableUI(id)
}
\arguments{
\item{id}{Namespace ID.}
}
\description{
A reusable UI component that displays a searchable, selectable gene table.
}
