% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scgeyser_utils.R
\name{get_core_plot_theme}
\alias{get_core_plot_theme}
\title{HELPER FUNCTIONS | R/utils.R ----
Define Core Plotting Theme}
\usage{
get_core_plot_theme()
}
\value{
A list of ggplot theme elements.
}
\description{
Creates a consistent ggplot theme for all plots to reduce code duplication.
}
