% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geyser.R
\name{geyser}
\alias{geyser}
\title{geyser}
\usage{
geyser(
  rse = NULL,
  app_name = "geyser",
  primary_color = "#3A5836",
  secondary_color = "#d5673e",
  computer_data_dir = NULL
)
}
\arguments{
\item{rse}{SummarizedExperiment object. If NULL (the default), the app will start with a file upload screen.}

\item{app_name}{Title name that goes on the top left of the Shiny app}

\item{primary_color}{The title bar color}

\item{secondary_color}{The plot action button color}

\item{computer_data_dir}{Optional folder path to existing SummarizedExperiment RDS files.}
}
\description{
Run shiny app to use SummarizedExperiment object to display genomics data
}
