% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghistApp.R
\name{gghistApp}
\alias{gghistApp}
\alias{gghistServer}
\alias{gghistInput}
\alias{gghistUI}
\alias{gghistOutput}
\title{Shiny App for Geyser Ggplot2 Histogram}
\usage{
gghistApp()

gghistServer(id, df = shiny::reactive(faithful))

gghistInput(id)

gghistUI(id)

gghistOutput(id)
}
\arguments{
\item{id}{shiny identifier}

\item{df}{reactive data frame}
}
\description{
Shiny App for Geyser Ggplot2 Histogram
}
