% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghist.R
\name{gghistServer}
\alias{gghistServer}
\alias{gghistInput}
\alias{gghistUI}
\alias{gghistOutput}
\alias{gghistApp}
\title{Shiny Server for Geyser Ggplot2 Histogram}
\usage{
gghistServer(id, df = shiny::reactive(faithful))

gghistInput(id)

gghistUI(id)

gghistOutput(id)

gghistApp()
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{df}{reactive data frame}
}
\value{
reactive server

nothing returned

nothing returned

nothing returned

nothing returned
}
\description{
Shiny Server for Geyser Ggplot2 Histogram

Shiny Module Input for Geyser Ggplot2 Histogram

Shiny Module UI for Geyser Ggplot2 Histogram

Shiny Module Output for Geyser Ggplot2 Histogram

Shiny Module App for Geyser Ggplot2 Histogram
}
