% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpointApp.R
\name{ggpointApp}
\alias{ggpointApp}
\alias{ggpointServer}
\alias{ggpointInput}
\alias{ggpointUI}
\alias{ggpointOutput}
\title{Shiny App for Geyser Ggplot2 Point Plot}
\usage{
ggpointApp()

ggpointServer(id, df = shiny::reactive(faithful))

ggpointInput(id)

ggpointUI(id)

ggpointOutput(id)
}
\arguments{
\item{id}{shiny identifier}

\item{df}{reactive data frame}
}
\description{
Shiny App for Geyser Ggplot2 Point Plot
}
