% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histApp.R
\name{histApp}
\alias{histApp}
\alias{histServer}
\alias{histInput}
\alias{histUI}
\alias{histOutput}
\title{Shiny App for Geyser Graphics Histogram}
\usage{
histApp()

histServer(id, df = shiny::reactive(faithful))

histInput(id)

histUI(id)

histOutput(id)
}
\arguments{
\item{id}{shiny identifier}

\item{df}{reactive data frame}
}
\description{
Shiny App for Geyser Graphics Histogram
}
