% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist.R
\name{histServer}
\alias{histServer}
\alias{histInput}
\alias{histUI}
\alias{histOutput}
\alias{histApp}
\title{Shiny Server for Geyser Graphics Histogram}
\usage{
histServer(id, df = shiny::reactive(faithful))

histInput(id)

histUI(id)

histOutput(id)

histApp()
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{df}{reactive data frame}
}
\value{
reactive server

nothing returned

nothing returned

nothing returned

nothing returned
}
\description{
Shiny Server for Geyser Graphics Histogram

Shiny Module Input for Geyser Graphics Histogram

Shiny Module UI for Geyser Graphics Histogram

Shiny Module Output for Geyser Graphics Histogram

Shiny Module App for Geyser Graphics Histogram
}
