% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowsApp.R
\name{rowsApp}
\alias{rowsApp}
\alias{rowsServer}
\alias{rowsInput}
\alias{rowsUI}
\title{Shiny App for Geyser Rows}
\usage{
rowsApp()

rowsServer(id)

rowsInput(id)

rowsUI(id)
}
\arguments{
\item{id}{shiny identifier}
}
\description{
Shiny App for Geyser Rows
}
