% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rows.R
\name{rowsServer}
\alias{rowsServer}
\alias{rowsInput}
\alias{rowsUI}
\alias{rowsApp}
\title{Shiny Server for Geyser Rows}
\usage{
rowsServer(id)

rowsInput(id)

rowsUI(id)

rowsApp()
}
\arguments{
\item{id}{shiny identifier}
}
\value{
reactive server
}
\description{
Shiny Server for Geyser Rows

Shiny Module Input

Shiny Module UI

Shiny Module App
}
