% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_helpers.R
\name{setup_config_upload_observer}
\alias{setup_config_upload_observer}
\title{Setup configuration upload observer}
\usage{
setup_config_upload_observer(input, session, rv)
}
\arguments{
\item{input}{Shiny input object}

\item{session}{Shiny session object}

\item{rv}{reactiveValues object}
}
\description{
Setup configuration upload observer
}
