% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switchApp.R
\name{switchApp}
\alias{switchApp}
\alias{switchServer}
\alias{switchInput}
\alias{switchUI}
\alias{switchOutput}
\title{Shiny App for Geyser Data Switch}
\usage{
switchApp()

switchServer(id)

switchInput(id)

switchUI(id)

switchOutput(id)
}
\arguments{
\item{id}{shiny identifier}
}
\description{
Shiny App for Geyser Data Switch
}
