% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch.R
\name{switchServer}
\alias{switchServer}
\alias{switchInput}
\alias{switchUI}
\alias{switchOutput}
\alias{switchApp}
\title{Shiny Server for Geyser Data Switch}
\usage{
switchServer(id)

switchInput(id)

switchUI(id)

switchOutput(id)

switchApp()
}
\arguments{
\item{id}{shiny identifier}
}
\value{
reactive server

nothing returned

nothing returned

nothing returned

nothing returned
}
\description{
Shiny Server for Geyser Data Switch

Shiny Module Input

Shiny Module UI

Shiny Module Output

Shiny Module App
}
