% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scgeyser_module_table_explorer.R
\name{tableExplorerControlsUI}
\alias{tableExplorerControlsUI}
\title{UI for Table Explorer Controls}
\usage{
tableExplorerControlsUI(id)
}
\arguments{
\item{id}{A namespace ID.}
}
\description{
Creates the UI for the table explorer controls in the sidebar.
}
