% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_table_explorer.R
\name{tableExplorerOutputUI}
\alias{tableExplorerOutputUI}
\title{UI for Table Explorer Output}
\usage{
tableExplorerOutputUI(id)
}
\arguments{
\item{id}{A namespace ID.}
}
\description{
UI for Table Explorer Output
}
