% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_table_explorer.R
\name{tableExplorerServer}
\alias{tableExplorerServer}
\title{Server for the Table Explorer Module}
\usage{
tableExplorerServer(id, loaded_data)
}
\arguments{
\item{id}{A namespace ID.}

\item{loaded_data}{A reactive list from the dataLoaderServer module.}
}
\description{
Server for the Table Explorer Module
}
