% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scgeyser_module_umap_plot.R
\name{umapPlotOutputUI}
\alias{umapPlotOutputUI}
\title{UI for UMAP Plot Output Area}
\usage{
umapPlotOutputUI(id)
}
\arguments{
\item{id}{Namespace ID.}
}
\description{
Creates the UI for the UMAP plot output in the main panel.
}
