% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_umap_plot.R
\name{umapPlotServer}
\alias{umapPlotServer}
\title{Server for the UMAP Plot Module}
\usage{
umapPlotServer(id, loaded_data)
}
\arguments{
\item{id}{Namespace ID.}

\item{loaded_data}{A reactive list from the dataLoaderServer module.}
}
\description{
Handles the logic for generating and displaying UMAP plots for gene expression and metadata.
}
