% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperApp.R
\name{wrapperApp}
\alias{wrapperApp}
\alias{wrapperServer}
\alias{wrapperInput}
\alias{wrapperUI}
\alias{wrapperOutput}
\title{Shiny App for Geyser Data Wrapper}
\usage{
wrapperApp()

wrapperServer(id)

wrapperInput(id)

wrapperUI(id)

wrapperOutput(id)
}
\arguments{
\item{id}{shiny identifier}
}
\description{
Shiny App for Geyser Data Wrapper
}
