% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{wrapperServer}
\alias{wrapperServer}
\alias{wrapperInput}
\alias{wrapperUI}
\alias{wrapperOutput}
\alias{wrapperApp}
\title{Shiny Server for Geyser Data Wrapper}
\usage{
wrapperServer(id)

wrapperInput(id)

wrapperUI(id)

wrapperOutput(id)

wrapperApp()
}
\arguments{
\item{id}{shiny identifier}
}
\value{
reactive server

nothing returned

nothing returned

nothing returned

nothing returned
}
\description{
Shiny Server for Geyser Data Wrapper

Shiny Module Input

Shiny Module UI

Shiny Module Output

Shiny Module App
}
