% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperset.R
\name{wrappersetServer}
\alias{wrappersetServer}
\alias{wrappersetInput}
\alias{wrappersetUI}
\alias{wrappersetOutput}
\alias{wrappersetApp}
\title{Shiny Server for Geyser Wrapper}
\usage{
wrappersetServer(id)

wrappersetInput(id)

wrappersetUI(id)

wrappersetOutput(id)

wrappersetApp()
}
\arguments{
\item{id}{shiny identifier}
}
\value{
reactive server

nothing returned

nothing returned

nothing returned

nothing returned
}
\description{
Shiny Server for Geyser Wrapper

Shiny Module Input

Shiny Module UI

Shiny Module Output

Shiny Module App
}
