% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data_loader.R
\name{dataLoaderUI}
\alias{dataLoaderUI}
\title{UI for the Data Loader Module (Sidebar Controls)}
\usage{
dataLoaderUI(id)
}
\arguments{
\item{id}{Namespace ID.}
}
\description{
Creates the UI for the data loader module, which appears in the sidebar.
}
