% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geyser_heatmap.R
\name{.hm_plot}
\alias{.hm_plot}
\title{hm_plot}
\usage{
.hm_plot(input, rse, slot)
}
\arguments{
\item{input}{From ui.R}

\item{rse}{rse object}

\item{slot}{which slot to pull the count data from the rse assay}
}
\value{
Returns a ComplexHeatmap object
}
\description{
draws the expression heatmap
}
\details{
Makes the heatmap for the geyser Shiny app
}
\examples{
load(system.file('extdata/tiny_rse.Rdata', package = 'geyser'))
input <- list()
input$feature_col <- 'row names'
input$features <- c("TYRP1 (ENSG00000107165.12)","OPN1LW (ENSG00000102076.9)")
input$groupings <- c('disease')
input$slot <- 'counts'
input$expression_scale <- TRUE
input$row_clust <- TRUE
input$col_clust <- TRUE
geyser:::.hm_plot(input, tiny_rse, 'counts')$plot
}
\author{
David McGaughey
}
\keyword{internal}
