% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geyser.R
\name{geyser}
\alias{geyser}
\title{geyser}
\usage{
geyser(rse, name = "geyser", ...)
}
\arguments{
\item{rse}{SummarizedExperiment object}

\item{name}{Title name that goes on the top left of the Shiny app}

\item{...}{More arguments for \code{\link[shiny:runApp]{shiny::runApp()}}.}
}
\description{
Run shiny app to use SummarizedExperiment object to display RNAseq data
}
\details{
Shiny app uses the rowData rownames to define the genes. The colData field is made
fully available to make custom plot groupings.
}
\examples{

\dontrun{
geyser(your_rse)
}

}
\author{
David McGaughey
}
