% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_table_explorer.R
\name{tableExplorerControlsUI}
\alias{tableExplorerControlsUI}
\title{UI for Table Explorer Controls}
\usage{
tableExplorerControlsUI(id)
}
\arguments{
\item{id}{A namespace ID.}
}
\description{
UI for Table Explorer Controls
}
