% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scgeyser_module_table_explorer.R
\name{tableExplorerOutputUI}
\alias{tableExplorerOutputUI}
\title{UI for Table Explorer Output}
\usage{
tableExplorerOutputUI(id)
}
\arguments{
\item{id}{A namespace ID.}
}
\description{
Creates the UI for the table explorer output in the main panel.
}
